move page loading code into loading.tsx
