+++ src/app/(default)/@modal/(.)work/[id]/loading.tsx
@@ -1,0 +1,3 @@
+import Loading from "@/app/(default)/work/[id]/loading"
+
+export default Loading
