+++ src/app/(default)/work/[id]/LayoutContent.module.css
@@ -16,22 +16,6 @@
   justify-content: center;
 }
 
-.loadingSkeletonBar {
-  width: 100%;
-  background-color: var(--gray4);
-
-  border-radius: 4px;
-}
-
-.loadingMockChildren {
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  padding-top: 96px;
-  padding-bottom: 144px;
-  gap: 2px;
-}
-
 .contentFadeIn {
   animation-name: moveUpFadeIn;
   animation-duration: 400ms;
@@ -122,11 +106,6 @@
   .contentFadeIn {
     flex-direction: column-reverse !important;
     gap: 16px;
-  }
-
-  .loadingMockChildren {
-    padding-left: 16px;
-    padding-right: 16px;
   }
 
   .leftSidebar {
