+++ src/app/(default)/work/[id]/loading.module.css
@@ -1,0 +1,40 @@
+.container {
+  z-index: 2;
+  background-color: var(--background);
+
+  display: flex;
+  flex-direction: row;
+
+  position: absolute;
+  left: 0;
+  right: 0;
+  top: 0;
+
+  min-height: 100vh;
+  width: 100vw;
+  flex: 1;
+  justify-content: center;
+}
+
+.loadingSkeletonBar {
+  width: 100%;
+  background-color: var(--gray4);
+
+  border-radius: 4px;
+}
+
+.loadingMockChildren {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  padding-top: 96px;
+  padding-bottom: 144px;
+  gap: 2px;
+}
+
+@media only screen and (max-width: 1200px) {
+  .loadingMockChildren {
+    padding-left: 16px;
+    padding-right: 16px;
+  }
+}
